flask_server/app/request_form/LoginForm.py
from flask_wtf import FlaskForm
from wtforms import StringField, PasswordField, SubmitField
from wtforms.validators import DataRequired

class LoginForm(FlaskForm):
    # Dulu namanya 'email', sekarang kita ganti jadi 'username'
    # Biar sinkron sama HTML dan Controller
    username = StringField('Username or Email', validators=[DataRequired()])
    password = PasswordField('Password', validators=[DataRequired()])
    submit = SubmitField('Login')